fixes with stock control

+++ api/Service/CartProductService.cs
@@ -7,23 +7,25 @@ namespace api.Service
 {
     public interface ICartProductService
     {
+        Task<CartProductStock> StockCheckAsync(int id, bool IsVariant);
         Task<bool> AddProductAsync(AddProductToCartDto dto);
-        Task<bool> RemoveProductAsync(int id);
-        Task<bool> UpdateProductQuantityAsync(int id, int quantity);
+        Task RemoveProductAsync(int id);
+        Task UpdateProductQuantityAsync(int id, int quantity, bool replinish);
     }
 
     public class CartProductService(
         AppDatabaseContext context,
         ICartRepository cartRepository,
         IUserService userService,
+        IProductRepository productRepository,
+        IProductVariantRepository productVariantRepository,
         ICartProductRepository cartProductRepository) : EFRepositoryBase<CartProduct>(context), ICartProductService
     {
         private readonly ICartRepository _cartRepository = cartRepository;
         private readonly IUserService _userService = userService;
         private readonly ICartProductRepository _cartProductRepository = cartProductRepository;
-
-
-
+        private readonly IProductRepository _productRepository = productRepository;
+        private readonly IProductVariantRepository _productVariantRepository = productVariantRepository;
 
         public async Task<bool> AddProductAsync(AddProductToCartDto dto)
         {
@@ -86,17 +88,74 @@ namespace api.Service
                     VariantId = dto.VariantId
                 });
 
+            if (create is { Data: not null })
+            {
+                if (!await UpdateStock(create.Data.Id, replinish: false, dto.Quantity))
+                {
+                    await _cartProductRepository.RemoveProductAsync(create.Data.Id);
+                    return false;
+                }
+            }
+
             return create.ErrorMsg is not null;
         }
 
-        public async Task<bool> RemoveProductAsync(int id)
+        public async Task RemoveProductAsync(int id)
         {
-            return await _cartProductRepository.RemoveProductAsync(id);
+            if (await UpdateStock(id, quantity: 0, replinish: true))
+            {
+                await _cartProductRepository.RemoveProductAsync(id);
+            }
         }
 
-        public async Task<bool> UpdateProductQuantityAsync(int id, int quantity)
+        public async Task UpdateProductQuantityAsync(int id, int quantity, bool replinish)
         {
-            return await _cartProductRepository.UpdateProductQuantityAsync(id, quantity);
+
+            if (await UpdateStock(id, replinish, quantity))
+            {
+                await _cartProductRepository.UpdateProductQuantityAsync(id, quantity);
+            }
+        }
+
+        public async Task<CartProductStock> StockCheckAsync(int id, bool IsVariant)
+        {
+            if (IsVariant)
+            {
+                return await _cartProductRepository.ProductVariantStockCheckAsync(id);
+            }
+
+            return await _cartProductRepository.ProductStockCheckAsync(id);
+        }
+
+        private async Task<bool> UpdateStock(int id, bool replinish, int quantity)
+        {
+            var cartProduct = await _cartProductRepository.GetAsync(id);
+
+            bool productIsVariant = cartProduct.VariantId.HasValue;
+            var stockCheck = await StockCheckAsync(cartProduct.VariantId ?? cartProduct.ProductId!.Value, productIsVariant);
+
+            if (stockCheck.Stock is null)
+                return false;
+
+            if (!replinish && stockCheck.Quantity > stockCheck.Stock)
+                return false;
+
+            var stock = replinish ?
+                stockCheck.Stock.Value + (quantity == 0 ? stockCheck.Quantity : quantity) :
+                stockCheck.Stock.Value - quantity;
+
+            stock = stock < 0 ? 0 : stock;
+
+            if (productIsVariant)
+            {
+                await _productVariantRepository.UpdateStockAsync(stockCheck.Id, stock);
+            }
+            else
+            {
+                await _productRepository.UpdateStockAsync(stockCheck.Id, stock);
+            }
+
+            return true;
         }
     }
 }
